Sending player input packet instead of heartbeat packet to ensure player synchronization

+++ Assets/Scripts/System/GameStateManager.cs
@@ -121,7 +121,7 @@ public class GameStateManager : MonoBehaviour
             state.PlayerIndex = player.PlayerIndex;
 
             PlayerInputData inputData = new PlayerInputData();
-            inputData.FrameNumber = (ushort)FrameCount;
+            inputData.FrameNumber = FrameCount;
             inputData.InputPattern = interpreter.GetPlayerInputByte();
         }
